+++ types/account.go
@@ -1,76 +1,1 @@
 package types
-
-import (
-	"fmt"
-	"math/big"
-
-	"github.com/umbracle/fastrlp"
-)
-
-// Account is an object we can retrieve from the state
-type Account struct {
-	Nonce    uint64
-	Balance  *big.Int
-	Root     Hash
-	CodeHash []byte
-}
-
-func (a *Account) Copy() *Account {
-	aa := new(Account)
-
-	aa.Balance = big.NewInt(1).SetBytes(a.Balance.Bytes())
-	aa.Nonce = a.Nonce
-	aa.CodeHash = a.CodeHash
-	aa.Root = a.Root
-
-	return aa
-}
-
-var accountParserPool fastrlp.ParserPool
-
-func (a *Account) UnmarshalRlp(b []byte) error {
-	p := accountParserPool.Get()
-	defer accountParserPool.Put(p)
-
-	v, err := p.Parse(b)
-	if err != nil {
-		return err
-	}
-	elems, err := v.GetElems()
-	if err != nil {
-		return err
-	}
-	if len(elems) != 4 {
-		return fmt.Errorf("bad")
-	}
-
-	// nonce
-	if a.Nonce, err = elems[0].GetUint64(); err != nil {
-		return err
-	}
-	// balance
-	if a.Balance == nil {
-		a.Balance = new(big.Int)
-	}
-	if err = elems[1].GetBigInt(a.Balance); err != nil {
-		return err
-	}
-	// root
-	if err = elems[2].GetHash(a.Root[:]); err != nil {
-		return err
-	}
-	// codeHash
-	if a.CodeHash, err = elems[3].GetBytes(a.CodeHash[:0]); err != nil {
-		return err
-	}
-	return nil
-}
-
-func (a *Account) MarshalWith(ar *fastrlp.Arena) *fastrlp.Value {
-	v := ar.NewArray()
-	v.Set(ar.NewUint(a.Nonce))
-	v.Set(ar.NewBigInt(a.Balance))
-	v.Set(ar.NewBytes(a.Root.Bytes()))
-	v.Set(ar.NewBytes(a.CodeHash))
-	return v
-}
